+++ annofabcli/statistics/visualization/dataframe/worktime_per_date.py
@@ -43,7 +43,18 @@ class WorktimePerDate:
     `date`,`account_id`がユニークなキーです。
     """
 
+    @staticmethod
+    def _duplicated_keys(df: pandas.DataFrame) -> bool:
+        """
+        DataFrameに重複したキーがあるかどうかを返します。
+        """
+        duplicated = df.duplicated(subset=["date", "account_id"])
+        return duplicated.any()
+
     def __init__(self, df: pandas.DataFrame) -> None:
+        if self._duplicated_keys(df):
+            logger.warning("引数`df`に重複したキー（date, account_id）が含まれています。")
+
         self.df = df
 
     _df_dtype = {
