+++ tests/data/statistics/task-worktime-by-user-phase.csv
@@ -1,0 +1,6 @@
+﻿project_id,task_id,status,phase,phase_stage,account_id,user_id,username,biography,worktime_hour,task_count,input_data_count,annotation_count,pointed_out_inspection_comment_count,rejected_count
+prj1,task1,complete,annotation,1,alice,alice,Alice,USA,2.0,1.0,2,100,5,1
+prj1,task1,complete,inspection,1,bob,bob,Bob,USA,0.5,0.5,1,50,0,0
+prj1,task2,complete,annotation,1,bob,bob,Bob,USA,2.0,1.0,4,200,6,1
+prj1,task2,complete,inspection,1,alice,alice,Alice,USA,0.5,0.5,2,100,0,0
+prj1,task2,complete,acceptance,1,alice,alice,Alice,USA,0.1,1.0,4,200,0,0
