+++ tests/data/statistics/task.csv
@@ -2,5 +2,5 @@ project_id,task_id,phase,phase_stage,status,user_id,username,number_of_rejection
 12345678-abcd-1234-abcd-1234abcd5678,task1,acceptance,1,complete,user3,user1,1,0,1,2019-11-25T04:18:05.8+09:00,2019-11-25T04:21:32.721+09:00,2019-11-25T04:21:32.56+09:00,,user1,user1,0.458071666666667,2019-11-21T22:35:32.238+09:00,user1,user1,0,2019-11-22T20:26:36.833+09:00,user1,user1,0.209477777777778,2019-11-22T20:26:36.833+09:00,0.787161388888889,0.520250277777778,1,0.266911111111111,10,41,9,False,True,,0.910469849537037,3.24028150462963
 12345678-abcd-1234-abcd-1234abcd5678,task2,acceptance,1,complete,user1,user2,2,0,2,2019-11-15T13:40:32.074+09:00,2019-11-15T13:43:35.02+09:00,2019-11-15T13:43:34.974+09:00,,user1,user1,0.801513055555556,2019-11-13T14:46:41.236+09:00,user2,user2,0,2019-11-15T10:12:26.528+09:00,user2,user2,0.350685555555556,2019-11-15T10:12:26.528+09:00,1.70834722222222,1.25867722222222,2,0.44967,10,65,7,True,False,,1.80955199074074,1.95617752314815
 12345678-abcd-1234-abcd-1234abcd5678,task10,acceptance,1,complete,user1,user4,0,0,0,2019-11-14T09:15:50.754+09:00,2019-11-14T09:18:49.676+09:00,2019-11-14T09:18:49.616+09:00,,user4,user4,0.217835277777778,2019-11-11T13:26:27.456+09:00,user4,user4,0,2019-11-14T09:15:50.754+09:00,user4,user4,0.049683888888889,2019-11-14T09:15:50.754+09:00,0.267519166666667,0.217835277777778,3,0.049683888888889,10,30,0,False,False,,2.82596409722222,2.82803425925926
-12345678-abcd-1234-abcd-1234abcd5678,task2,acceptance,1,not_started,,,0,0,0,2019-11-25T22:05:23.501+09:00,2019-11-25T22:09:02.157+09:00,,,user4,user4,0.060726388888889,2019-11-25T22:05:23.501+09:00,,,0,,,,0,,0.060726388888889,0.060726388888889,0,0,10,10,0,False,False,,,
+12345678-abcd-1234-abcd-1234abcd5678,task4,acceptance,1,not_started,,,0,0,0,2019-11-25T22:05:23.501+09:00,2019-11-25T22:09:02.157+09:00,,,user4,user4,0.060726388888889,2019-11-25T22:05:23.501+09:00,,,0,,,,0,,0.060726388888889,0.060726388888889,0,0,10,10,0,False,False,,,
 12345678-abcd-1234-abcd-1234abcd5678,task5,annotation,1,not_started,,,0,0,0,,2019-11-08T22:54:26.981+09:00,,,user5,user5,0.029444166666667,2019-11-26T09:27:22.776+09:00,,,0,,,,0,,0.029444166666667,0.029444166666667,0,0,10,0,0,False,False,,,
